add cities empty component
